Move config rendering logic to new RenderConfigMixin
